remove vault index (replace with global index)

+++ contracts/interfaces/vault/IVaultRestricted.sol
@@ -9,7 +9,7 @@ interface IVaultRestricted {
         address[] calldata vaultStrategies,
         uint256 newVaultProportions,
         uint256 finishedIndex,
-        uint256 activeIndex
+        uint24 activeIndex
     ) external returns (uint256[] memory, uint256);
 
     function payFees(uint256 profit) external returns (uint256 feesPaid);
